refactor: move input-agnostic helpers to utils/input
The `defineSubmitAnswerExternalAction` and
`returnInputValueChangedStateVariableDefinitions` helpers are used by
non-math inputs (textInput, codeEditor, booleanInput, choiceInput) as
well as the math inputs, so they no longer belong in the math-specific
`utils/mathComponentInput` module. Relocate them to a new input-agnostic
`utils/input` module and update all consumers.
