+++ packages/doenetml-worker-javascript/src/components/MatrixInput.js
@@ -15,15 +15,17 @@ import {
 } from "../utils/numberDisplay";
 import { roundForDisplay } from "../utils/math";
 import {
-    defineSubmitAnswerExternalAction,
     mathComponentInputFocusChanged,
     mathComponentInputUpdateRawValue,
     mathComponentInputUpdateValue,
-    returnInputValueChangedStateVariableDefinitions,
     returnMathComponentInputConfigStateVariableDefinitions,
     returnMathComponentInputDisplayStateVariableDefinitions,
     returnMathInputParsingAttributes,
 } from "../utils/mathComponentInput";
+import {
+    defineSubmitAnswerExternalAction,
+    returnInputValueChangedStateVariableDefinitions,
+} from "../utils/input";
 
 const vectorAndListOperators = ["list", ...vectorOperators];
