+++ packages/doenetml-worker-javascript/src/utils/mathComponentInput.js
@@ -21,30 +21,6 @@ import {
  */
 
 /**
- * Define a `submitAnswer` external action on an input component that delegates
- * to its `answerAncestor`. Call from the component constructor.
- */
-export function defineSubmitAnswerExternalAction(component) {
-    component.externalActions = {};
-
-    // Complex because the stateValues isn't defined until later
-    Object.defineProperty(component.externalActions, "submitAnswer", {
-        enumerable: true,
-        get: async function () {
-            let answerAncestor = await this.stateValues.answerAncestor;
-            if (answerAncestor !== null) {
-                return {
-                    componentIdx: answerAncestor.componentIdx,
-                    actionName: "submitAnswer",
-                };
-            } else {
-                return;
-            }
-        }.bind(component),
-    });
-}
-
-/**
  * The parsing attributes (`format`, `functionSymbols`, `splitSymbols`,
  * `parseScientificNotation`) shared by inputs that parse math the same way a
  * `<mathInput>` does.
@@ -93,73 +69,6 @@ export function returnMathInputParsingAttributes() {
             public: true,
         },
     };
-}
-
-/**
- * The `valueChanged` / `immediateValueChanged` essential booleans, public so
- * authors can detect whether a user has edited the input.
- */
-export function returnInputValueChangedStateVariableDefinitions({
-    valueChangedDescription,
-    immediateValueChangedDescription,
-} = {}) {
-    const stateVariableDefinitions = {};
-
-    stateVariableDefinitions.valueChanged = {
-        description: valueChangedDescription,
-        public: true,
-        hasEssential: true,
-        defaultValue: false,
-        shadowingInstructions: {
-            createComponentOfType: "boolean",
-        },
-        returnDependencies: () => ({}),
-        definition() {
-            return { useEssentialOrDefaultValue: { valueChanged: true } };
-        },
-        inverseDefinition({ desiredStateVariableValues }) {
-            return {
-                success: true,
-                instructions: [
-                    {
-                        setEssentialValue: "valueChanged",
-                        value: Boolean(desiredStateVariableValues.valueChanged),
-                    },
-                ],
-            };
-        },
-    };
-
-    stateVariableDefinitions.immediateValueChanged = {
-        description: immediateValueChangedDescription,
-        public: true,
-        hasEssential: true,
-        defaultValue: false,
-        shadowingInstructions: {
-            createComponentOfType: "boolean",
-        },
-        returnDependencies: () => ({}),
-        definition() {
-            return {
-                useEssentialOrDefaultValue: { immediateValueChanged: true },
-            };
-        },
-        inverseDefinition({ desiredStateVariableValues }) {
-            return {
-                success: true,
-                instructions: [
-                    {
-                        setEssentialValue: "immediateValueChanged",
-                        value: Boolean(
-                            desiredStateVariableValues.immediateValueChanged,
-                        ),
-                    },
-                ],
-            };
-        },
-    };
-
-    return stateVariableDefinitions;
 }
 
 /**
